Profile creation implementation (#6)
The profile in the database is now created through the form on the site.

After creating a profile, we are redirected to a page that counts from 20 to zero and automatically redirects us to the newly created profile. We also have the opportunity to go to it manually by pressing the transition button.