Product page (#15)
* backup

* backup

Made drawing an indefinite number of thumbnails when you click on one of which it is shown in the main window

* task_completed

The product page is ready and working in test mode, displaying only one manually written product in the code. In order to display the products recorded in the database in the next branch, I need to add tables that will store these products in the database.