Products in the basket (#18)
* backup_before_editing_database_tables

* backup

* updating_the_database_of_objects_and_images_for_them

* backup

* backup

Now you can get to the product page by the usual method - choosing a product. No need to manually enter the address anymore.

* site_visual_bug_fixes
backup_before_reconstructing_product_page

* backup

* backup_2

* product_page_finished

The technical part on the product page is finished.

* backup_3

Backup before redesigning the recycle bin window a bit

* items_now_line_up_properly_on_the_cart_page

* auto-loading_links

Now the product is loaded immediately when called, along with links to images.

* backup_4

Backup before moving the cart methods to the new controller

* set_up_the_filter_HiddenHttpMethodFilter

* removing_product

Implemented removal of a product from the database

* сhanging_the_quantity_of_goods

Now you can add and subtract the number of items for a specific item in your cart.

* processing_multiple_products_in_a_shopping_cart

Now you can normally add and subtract the number of products for bundles in the basket when there are more than one bundles. Before that, the site could correctly process such operations for only one bundle.

* backup_before_changing_the_database_structure

* bug_fix

Removed a new bug because of which the user could not register on the site

* visual_edits